Seperate routing server socket code
